Move getGlobalStorage() to where similar helpers are.
